Refactor: move TokenError to own error.rs, remove parser.rs
Moves the statements of parser.rs into the lexer.rs and moves TokenError
into own error crate to make it easier to write the actual parser
without bias and have a better structure to work with.
